Refactor React routes into their own file
