+++ app/queries/Organization.js
@@ -5,3 +5,5 @@ export const query = () => Relay.QL`
     organization(slug: $organization)
   }
 `;
+
+export default { query };
