+++ app/queries/Pipeline.js
@@ -12,3 +12,5 @@ export const prepareParams = (params) => {
     slug: [params.organization, params.pipeline].join("/")
   };
 };
+
+export default { query, prepareParams };
